Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,9 +3,9 @@
 
 
 int main() {
-    int a,b,c;
+    float a,b,c;
     float res;
-    scanf("%d%d%d" ,&a,&b,&c);
+    scanf("%f%f%f" ,&a,&b,&c);
     res=(a+b+c)/3;
     printf("Average: %.2f",res);
